--- tests/ORM/Functional/Driver/Common/Integration/Case4/Entity/PostTag.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case4\Entity;
-
-class PostTag
-{
-    private ?int $id = null;
-    private ?int $post_id = null;
-    private ?int $tag_id = null;
-}
