+++ tests/ORM/Functional/Driver/MySQL/Integration/Issue322/CaseTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ORM\Tests\Functional\Driver\MySQL\Integration\Case5;
+namespace Cycle\ORM\Tests\Functional\Driver\MySQL\Integration\Issue322;
 
 // phpcs:ignore
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case5\CaseTest as CommonClass;
+use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Issue322\CaseTest as CommonClass;
 
 /**
  * @group driver
