--- tests/ORM/Functional/Driver/SQLServer/Integration/Case4/CaseTest.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Functional\Driver\SQLServer\Integration\Case4;
-
-// phpcs:ignore
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case4\CaseTest as CommonClass;
-
-/**
- * @group driver
- * @group driver-sqlserver
- */
-class CaseTest extends CommonClass
-{
-    public const DRIVER = 'sqlserver';
-}
