--- tests/ORM/Functional/Driver/SQLite/Integration/Case4/CaseTest.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Functional\Driver\SQLite\Integration\Case4;
-
-// phpcs:ignore
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case4\CaseTest as CommonClass;
-
-/**
- * @group driver
- * @group driver-sqlite
- */
-class CaseTest extends CommonClass
-{
-    public const DRIVER = 'sqlite';
-}
